# apps/dsa-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DSA API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Health check"
          }
        }
      }
    },
    "/datasets": {
      "get": {
        "summary": "List datasets",
        "operationId": "listDatasets",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dataset"
                  }
                }
              }
            },
            "description": "List datasets"
          }
        }
      },
      "post": {
        "summary": "Upload dataset",
        "operationId": "uploadDataset",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            },
            "description": "Dataset created (5MB limit, MIME double-check)"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Payload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid file type"
          }
        }
      }
    },
    "/datasets/{id}": {
      "get": {
        "summary": "Get dataset",
        "operationId": "getDataset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetProfile"
                }
              }
            },
            "description": "Dataset profile (schema/preview<=20 rows/analyses/lineage)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          }
        }
      }
    },
    "/analysis": {
      "get": {
        "summary": "List analyses",
        "operationId": "listAnalyses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRun"
                  }
                }
              }
            },
            "description": "List analysis runs"
          }
        }
      },
      "post": {
        "summary": "Run analysis",
        "operationId": "createAnalysis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRunCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            },
            "description": "Analysis run created (includes runId)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Validation error"
          }
        }
      }
    },
    "/analysis/{runId}": {
      "get": {
        "summary": "Get analysis run",
        "operationId": "getAnalysisRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Full run detail (plan/trace/evidence/insights/limitations/artifacts/report_markdown)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          }
        }
      }
    },
    "/benchmarks": {
      "get": {
        "summary": "List benchmarks",
        "operationId": "listBenchmarks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenchmarkRow"
                  }
                }
              }
            },
            "description": "List benchmark rows"
          }
        }
      }
    },
    "/reports": {
      "get": {
        "summary": "List reports",
        "operationId": "listReports",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRun"
                  }
                }
              }
            },
            "description": "List reports"
          }
        }
      }
    },
    "/runs": {
      "get": {
        "summary": "List runs",
        "operationId": "listRuns",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRun"
                  }
                }
              }
            },
            "description": "List runs"
          }
        }
      }
    },
    "/runs/{id}": {
      "get": {
        "summary": "Get run",
        "operationId": "getRun",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Run detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          }
        }
      }
    },
    "/runs/{id}/replay": {
      "get": {
        "summary": "Replay run",
        "operationId": "getReplay",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Replay payload"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Dataset": {
        "type": "object",
        "required": [
          "id",
          "file",
          "format",
          "rows",
          "cols",
          "created"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "file": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "rows": {
            "type": "integer",
            "minimum": 0
          },
          "cols": {
            "type": "integer",
            "minimum": 0
          },
          "created": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          }
        }
      },
      "DatasetProfile": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Dataset"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "type"
                  ],
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  }
                }
              },
              "preview": {
                "type": "array",
                "maxItems": 20,
                "items": {
                  "type": "object"
                }
              },
              "analyses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              },
              "lineage": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "step",
                    "detail"
                  ],
                  "properties": {
                    "step": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "AnalysisRun": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "dataset": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created": {
            "type": "string"
          }
        }
      },
      "AnalysisRunCreate": {
        "type": "object",
        "required": [
          "dataset",
          "query"
        ],
        "additionalProperties": false,
        "properties": {
          "dataset": {
            "type": "string"
          },
          "query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
      "TraceStep": {
        "type": "object",
        "required": [
          "tool",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "tool": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "duration_ms": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "call_id": {
            "type": "string"
          },
          "input": {},
          "output": {}
        }
      },
      "EvidenceItem": {
        "type": "object",
        "required": [
          "claim",
          "confidence"
        ],
        "additionalProperties": false,
        "properties": {
          "claim": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 2
          },
          "source": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "result": {}
        }
      },
      "RunDetail": {
        "type": "object",
        "required": [
          "id",
          "objective",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "plan": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "step",
                "at"
              ],
              "properties": {
                "step": {
                  "type": "string"
                },
                "at": {
                  "type": "string"
                }
              }
            }
          },
          "validation_rate": {
            "type": "number"
          },
          "trace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            }
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            }
          },
          "insights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "limitations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "artifacts": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "name",
                "kind"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "kind": {
                  "type": "string"
                }
              }
            }
          },
          "report_markdown": {
            "type": "string"
          }
        }
      },
      "BenchmarkRow": {
        "type": "object",
        "required": [
          "task",
          "v1",
          "v2"
        ],
        "additionalProperties": false,
        "properties": {
          "task": {
            "type": "string"
          },
          "v1": {
            "type": "number"
          },
          "v2": {
            "type": "number"
          },
          "category": {
            "type": "string"
          }
        }
      },
      "Health": {
        "type": "object",
        "required": [
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "detail"
        ],
        "additionalProperties": false,
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Server-generated request id, echoed on every response (success and error).",
        "schema": {
          "type": "string"
        }
      }
    }
  }
}